Changed return type of getEntityCount() from long to int.
git-svn-id: http://dev.vaadin.com/svn/addons/JPAContainer/trunk@10279 de684a11-831f-0410-ab27-cb86445eb682

+++ src/main/java/com/vaadin/addons/jpacontainer/EntityProvider.java
@@ -111,5 +111,5 @@ public interface EntityProvider<T> {
      * @param filter the filter that should be used to filter the entities (may be null).
      * @return the number of matches.
      */
-    public long getEntityCount(Filter filter);
+    public int getEntityCount(Filter filter);
 }
